chore(security): remove the legacy prompt-risk filter
`atlas/core/prompt_risk.py` scored RAG chunks against prompt-injection
heuristics and appended medium/high hits to `logs/security_high_risk.jsonl`.
It never blocked or redacted anything, nothing in the tree read the log, and
no external collector consumed it. Its only call site was inside
`RAGMCPService.search_raw`, which the live path does not reach --
`UnifiedRAGService` calls `synthesize`, and `search_raw` runs only as an
internal fallback for MCP servers lacking `rag_get_synthesized_results`.

The hook system (#713 / #803) supersedes it: `RagCall` and `RagResponse` fire
around every retrieval, on both the single and batch paths and the agentic
`atlas_rag_query` path, and can block, rewrite the query, narrow sources, or
redact content rather than only observe.

Removed:
- `atlas/core/prompt_risk.py`
- its call site and import in `atlas/domain/rag_mcp_service.py`
- `pi_threshold_low/medium/high` settings, already marked "NOT USED RIGHT NOW"
  (`AppSettings` sets `extra: ignore`, so a stale `PI_THRESHOLD_*` in a
  deployment `.env` is ignored rather than an error)
- the two prompt-risk test classes in `tests/test_isolation_fixtures.py` and
  the prompt-risk tests in `tests/test_prompt_risk_and_acl.py`, whose
  surviving ACL test moves to `tests/test_tool_acl.py`

Kept, as opt-in operator policy: the scoring heuristics ship as an example
`RagResponse` hook, `atlas/config/hooks-example/rag_injection_scan.py`. It
observes only as written, logs where `$ATLAS_RAG_SCAN_LOG` points, and
documents how to turn it into an enforcement point.

This also supersedes #818's `APP_LOG_DIR` handling for
`security_high_risk.jsonl`, since the file is no longer written.
