--- atlas/core/prompt_risk.py
@@ -1,301 +1,0 @@
-"""
-Prompt injection risk heuristics and structured logging.
-
-Scope: lightweight, configurable thresholds; used for user input and RAG results.
-"""
-
-from __future__ import annotations
-
-import base64
-import json
-import logging
-import math
-import os
-import re
-from collections import Counter
-from datetime import datetime, timezone
-from pathlib import Path
-from typing import Dict, List, Optional
-
-logger = logging.getLogger(__name__)
-
-
-def _get_thresholds() -> Dict[str, int]:
-    """Get prompt injection risk thresholds from config manager."""
-    try:
-        from atlas.modules.config import config_manager
-        settings = config_manager.app_settings
-        return {
-            "low": settings.pi_threshold_low,
-            "medium": settings.pi_threshold_medium,
-            "high": settings.pi_threshold_high,
-        }
-    except Exception:
-        # Fallback to defaults if config not available
-        return {
-            "low": 30,
-            "medium": 50,
-            "high": 80,
-        }
-
-
-def calculate_prompt_injection_risk(message: str, *, mode: str = "general") -> Dict[str, object]:
-    """
-    Calculate a heuristic risk score for prompt injection attempts.
-
-    Returns: { 'score': int, 'risk_level': str, 'triggers': list[str] }
-    """
-    score = 0
-    triggers: List[str] = []
-
-    msg_lower = (message or "").lower()
-
-    # 1) Suspicious patterns
-    patterns = {
-        "override_instructions": (r"ignore\s+(previous|all|everything|above|prior)", 40),
-        "disregard": (r"disregard\s+(previous|all|everything|above|prior)", 40),
-        "new_instructions": (r"new\s+instructions?\s*:\s*", 35),
-        "system_role": (r"\b(system|assistant|user)\s*:\s*", 30),
-        "act_as": (r"act\s+as\s+(if\s+)?you\s+(are|were)", 25),
-        "pretend": (r"pretend\s+(to\s+be|you\s+are)", 25),
-        "role_change": (r"your?\s+(new\s+)?role\s+(is|now)", 30),
-        "forget": (r"forget\s+(everything|all|previous)", 35),
-        "override": (r"override\s+(previous|default|system)", 35),
-        "jailbreak": (r"(jailbreak|developer\s+mode|god\s+mode)", 45),
-    }
-    for name, (pat, pts) in patterns.items():
-        if re.search(pat, msg_lower):
-            score += pts
-            triggers.append(name)
-
-    # 2) Encodings/obfuscation
-    if _detect_encoding(message):
-        score += 30
-        triggers.append("encoding_detected")
-
-    # 3) Statistical anomalies
-    # Delimiter density (triple quotes, fences, etc.)
-    delimiters = len(re.findall(r"[#*\-_=]{3,}|[\"']{3,}", message or ""))
-    if delimiters >= 3:
-        score += 25
-        triggers.append("excessive_delimiters")
-    elif delimiters >= 1:
-        score += 10
-
-    # High entropy (possible encoded blob)
-    if len(message or "") > 10:
-        ent = _calculate_entropy(message)
-        if ent > 4.5:
-            score += 20
-            triggers.append("high_entropy")
-
-    # Excessive caps
-    if len(message or "") > 20:
-        caps_ratio = sum(1 for c in (message or "") if c.isupper()) / max(1, len(message or ""))
-        if caps_ratio > 0.3:
-            score += 15
-            triggers.append("excessive_caps")
-
-    # 4) Context-breaking attempts
-    if (message or "").count("\n") > 5 or re.search(r"\s{10,}", message or ""):
-        score += 15
-        triggers.append("formatting_abuse")
-
-    if re.search(r"(human|user|assistant):\s*\n", msg_lower):
-        score += 25
-        triggers.append("fake_conversation")
-
-    if (len(message or "") > 50) and re.search(r"<[^>]+>.*</[^>]+>|[{}\[\]]", message or ""):
-        score += 20
-        triggers.append("structured_injection")
-
-    # 5) Length penalty
-    if len(message or "") > 1000:
-        score += 15
-        triggers.append("excessive_length")
-
-    # Context-aware normalization (reduce false positives)
-    if mode in ("code", "logs"):
-        # Code/logs: braces, fences common; soften penalties
-        if "excessive_delimiters" in triggers:
-            score -= 10
-        if "structured_injection" in triggers:
-            score -= 10
-        score = max(0, score)
-
-    # Risk buckets - get thresholds from config
-    thresholds = _get_thresholds()
-    if score >= thresholds["high"]:
-        level = "high"
-    elif score >= thresholds["medium"]:
-        level = "medium"
-    elif score >= thresholds["low"]:
-        level = "low"
-    else:
-        level = "minimal"
-
-    return {"score": int(score), "risk_level": level, "triggers": triggers}
-
-
-def _detect_encoding(text: str) -> bool:
-    clean = re.sub(r"\s", "", text or "")
-    # Base64-like
-    if len(clean) > 20 and len(clean) % 4 == 0 and re.match(r"^[A-Za-z0-9+/=]+$", clean or ""):
-        try:
-            base64.b64decode(clean, validate=True)
-            return True
-        except Exception:
-            pass
-    # Hex
-    if re.match(r"^(0x)?[0-9a-fA-F]+$", clean or "") and len(clean) > 20:
-        return True
-    # Escape sequences
-    if re.search(r"\\u[0-9a-fA-F]{4}|\\x[0-9a-fA-F]{2}", text or ""):
-        return True
-    # Zero-width/unusual unicode
-    if re.search(r"[\u200B-\u200D\uFEFF\u2060]", text or ""):
-        return True
-    return False
-
-
-def _calculate_entropy(text: str) -> float:
-    if not text:
-        return 0.0
-    counts = Counter(text)
-    n = len(text)
-    ent = 0.0
-    for c in counts.values():
-        p = c / n
-        ent -= p * math.log2(max(p, 1e-12))
-    return ent
-
-
-# Paths already warned about, so a persistently unwritable log cannot flood the
-# very log it is meant to make legible. Keyed by path: a location that is fixed
-# and later breaks again warns again.
-_WRITE_FAILURE_WARNED_PATHS: set = set()
-
-# Resolved paths the relocation notice has already been emitted for. A set
-# rather than a rebound global: same reason as above, and it means a runtime
-# change of APP_LOG_DIR gets its own notice instead of being swallowed.
-_RELOCATION_NOTICE_PATHS: set = set()
-
-
-def high_risk_log_path() -> Path:
-    """Return the path of the high-risk security log.
-
-    Resolution is ``APP_LOG_DIR`` env var, then ``app_settings.app_log_dir``,
-    then the repository's ``logs/`` directory -- the same order as
-    ``routes/telemetry_routes._log_base_dir``. (``routes/admin_routes._log_base_dir``
-    consults only the setting and ``core/telemetry._tool_output_dir`` only the
-    env var; the union of the two is what a deployment actually expects, so
-    that is what this uses. Unifying all four is worth doing, but it is a
-    behavior change for those call sites and does not belong in this change.)
-
-    Honoring the override is also what keeps a test run, which points
-    ``APP_LOG_DIR`` at a temp directory, from appending to -- or truncating --
-    a real deployment's security log.
-    """
-    override = os.getenv("APP_LOG_DIR")
-    if not override:
-        try:
-            from atlas.modules.config import config_manager
-
-            override = config_manager.app_settings.app_log_dir
-        except Exception:  # pragma: no cover - config unavailable; use the default
-            override = None
-    base = Path(override) if override else _default_log_dir()
-    # expanduser/resolve so a ``~``-prefixed or CWD-relative configured value
-    # lands somewhere predictable rather than following the process's cwd.
-    path = (base.expanduser().resolve() / "security_high_risk.jsonl")
-    _warn_once_if_relocated(path)
-    return path
-
-
-def _default_log_dir() -> Path:
-    return Path(__file__).resolve().parents[2] / "logs"
-
-
-def _warn_once_if_relocated(path: Path) -> None:
-    """Say so, once, when a stale copy of this log exists at the old location.
-
-    Before this log honored ``APP_LOG_DIR`` it always lived in the repository's
-    ``logs/``. A collector still tailing that file would go quiet and read as
-    "no attacks observed", so name both paths -- but only when the old file
-    actually exists, so deployments that never had one stay silent.
-    """
-    if str(path) in _RELOCATION_NOTICE_PATHS:
-        return
-    _RELOCATION_NOTICE_PATHS.add(str(path))
-    try:
-        legacy = _default_log_dir() / "security_high_risk.jsonl"
-        if path != legacy and legacy.exists():
-            logger.warning(
-                "High-risk security log now writes to %s (APP_LOG_DIR); a "
-                "stale copy remains at %s and is no longer updated -- point "
-                "any collector at the new path.",
-                path, legacy,
-            )
-    except Exception:  # pragma: no cover - a notice must never break logging
-        pass
-
-
-def log_high_risk_event(*, source: str, user: Optional[str], content: str, score: int, risk_level: str, triggers: List[str], extra: Optional[Dict[str, object]] = None) -> None:
-    """Append a JSONL record for medium/high events to the high-risk log.
-
-    See ``high_risk_log_path`` for how the location is resolved.
-    """
-    log_path: Optional[Path] = None
-    try:
-        # Only log medium/high
-        if risk_level not in ("medium", "high"):
-            return
-        log_path = high_risk_log_path()
-        log_path.parent.mkdir(parents=True, exist_ok=True)
-        record = {
-            "ts": datetime.now(timezone.utc).isoformat() + "Z",
-            "type": "prompt_risk",
-            "source": source,
-            "user": user,
-            "score": score,
-            "risk_level": risk_level,
-            "triggers": triggers,
-        }
-        if extra:
-            record.update(extra)
-        # include a small snippet only
-        snippet = (content or "")
-        if len(snippet) > 240:
-            snippet = snippet[:240] + "…"
-        record["snippet"] = snippet
-        with open(log_path, "a", encoding="utf-8") as f:
-            f.write(json.dumps(record, ensure_ascii=False) + "\n")
-        # Writing again means the location is healthy; drop its key so a path
-        # that breaks a second time warns a second time. Without this the
-        # "repaired and later broken warns again" behavior below is a lie.
-        _WRITE_FAILURE_WARNED_PATHS.discard(str(log_path))
-    except Exception as e:
-        # A security audit record that cannot be written is an operational
-        # problem: without a signal, a misconfigured or unwritable path is
-        # indistinguishable from "no attacks observed". Warn once (this runs
-        # per event, and a persistently bad path would otherwise flood the
-        # log), then fall back to debug.
-        # Key on the resolved path, or -- when the resolver itself raised, so
-        # there is no path -- on the configured value that produced the
-        # failure. A bare "<unresolved>" constant would collapse every
-        # misconfiguration onto one key, so the second bad APP_LOG_DIR would
-        # only ever be logged at debug.
-        if log_path is not None:
-            failed_path = str(log_path)
-        else:
-            configured = os.getenv("APP_LOG_DIR") or "<unset>"
-            failed_path = f"<unresolved:{configured}>"
-        if failed_path not in _WRITE_FAILURE_WARNED_PATHS:
-            _WRITE_FAILURE_WARNED_PATHS.add(failed_path)
-            logger.warning(
-                "Failed to write high risk security log at %s: %s "
-                "(further failures for this path logged at debug)",
-                failed_path, e,
-            )
-        else:
-            logger.debug("Failed to write high risk log: %s", e)
