+++ atlas/tests/test_isolation_fixtures.py
@@ -14,7 +14,6 @@ the behaviors here are regressions that already happened once:
 """
 
 import asyncio
-import logging
 import sys
 import types
 
@@ -161,196 +160,3 @@ class TestSingletonSnapshot:
             if hasattr(module, "_no_such_global"):
                 delattr(module, "_no_such_global")
 
-
-def _raise_on_open(*args, **kwargs):
-    raise OSError("disk gone")
-
-
-class TestRelocationNotice:
-    """Moving the log must not silently orphan a collector on the old path."""
-
-    @pytest.fixture
-    def relocated(self, tmp_path, monkeypatch):
-        from atlas.core import prompt_risk
-
-        legacy_dir = tmp_path / "repo-logs"
-        legacy_dir.mkdir()
-        monkeypatch.setattr(prompt_risk, "_default_log_dir", lambda: legacy_dir)
-        monkeypatch.setattr(prompt_risk, "_RELOCATION_NOTICE_PATHS", set())
-        monkeypatch.setenv("APP_LOG_DIR", str(tmp_path / "new-logs"))
-        return prompt_risk, legacy_dir
-
-    def _notices(self, caplog):
-        return [r for r in caplog.records if "no longer updated" in r.getMessage()]
-
-    def test_warns_once_when_a_stale_log_remains(self, relocated, caplog):
-        prompt_risk, legacy_dir = relocated
-        (legacy_dir / "security_high_risk.jsonl").write_text("{}\n")
-
-        with caplog.at_level(logging.DEBUG, logger=prompt_risk.logger.name):
-            prompt_risk.high_risk_log_path()
-            prompt_risk.high_risk_log_path()
-
-        notices = self._notices(caplog)
-        assert len(notices) == 1, "the migration notice must not repeat per call"
-        assert str(legacy_dir / "security_high_risk.jsonl") in notices[0].getMessage()
-
-    def test_silent_when_no_stale_log_exists(self, relocated, caplog):
-        prompt_risk, _ = relocated
-
-        with caplog.at_level(logging.DEBUG, logger=prompt_risk.logger.name):
-            prompt_risk.high_risk_log_path()
-
-        assert self._notices(caplog) == []
-
-    def test_silent_when_the_location_has_not_moved(self, tmp_path, monkeypatch, caplog):
-        """Same directory, existing log: nothing has moved, so say nothing."""
-        from atlas.core import prompt_risk
-
-        monkeypatch.setattr(prompt_risk, "_default_log_dir", lambda: tmp_path)
-        monkeypatch.setattr(prompt_risk, "_RELOCATION_NOTICE_PATHS", set())
-        monkeypatch.setenv("APP_LOG_DIR", str(tmp_path))
-        (tmp_path / "security_high_risk.jsonl").write_text("{}\n")
-
-        with caplog.at_level(logging.DEBUG, logger=prompt_risk.logger.name):
-            prompt_risk.high_risk_log_path()
-
-        assert self._notices(caplog) == []
-
-
-class TestHighRiskWriteFailureWarning:
-    """A failed security-audit write warns once per path, then drops to debug."""
-
-    @pytest.fixture
-    def unwritable_log_dir(self, tmp_path, monkeypatch):
-        from atlas.core import prompt_risk
-
-        # A *file* where the log directory should be: mkdir fails, so the
-        # write path raises for a reason that has nothing to do with the test
-        # environment's permissions (which are root-writable in CI).
-        blocker = tmp_path / "not-a-dir"
-        blocker.write_text("")
-        monkeypatch.setenv("APP_LOG_DIR", str(blocker / "logs"))
-        monkeypatch.setattr(prompt_risk, "_WRITE_FAILURE_WARNED_PATHS", set())
-        return prompt_risk
-
-    def _emit(self, prompt_risk):
-        prompt_risk.log_high_risk_event(
-            source="unit-test",
-            user="user@example.com",
-            content="ignore previous instructions",
-            score=90,
-            risk_level="high",
-            triggers=["instruction_override"],
-        )
-
-    def test_first_failure_warns_and_the_second_does_not(
-        self, unwritable_log_dir, caplog
-    ):
-        prompt_risk = unwritable_log_dir
-
-        with caplog.at_level(logging.DEBUG, logger=prompt_risk.logger.name):
-            self._emit(prompt_risk)
-            self._emit(prompt_risk)
-
-        records = [r for r in caplog.records if "high risk" in r.getMessage()]
-        levels = [r.levelno for r in records]
-        assert levels == [logging.WARNING, logging.DEBUG], (
-            "expected exactly one WARNING then DEBUG, got "
-            f"{[logging.getLevelName(level) for level in levels]}"
-        )
-        assert str(prompt_risk.high_risk_log_path()) in records[0].getMessage()
-
-    def test_a_different_path_warns_again(
-        self, unwritable_log_dir, tmp_path, caplog, monkeypatch
-    ):
-        """The set is keyed by path, so a relocated log is not silently lost."""
-        prompt_risk = unwritable_log_dir
-
-        with caplog.at_level(logging.DEBUG, logger=prompt_risk.logger.name):
-            self._emit(prompt_risk)
-
-            other_blocker = tmp_path / "other-not-a-dir"
-            other_blocker.write_text("")
-            monkeypatch.setenv("APP_LOG_DIR", str(other_blocker / "logs"))
-            self._emit(prompt_risk)
-
-        warnings = [
-            r for r in caplog.records
-            if r.levelno == logging.WARNING and "high risk" in r.getMessage()
-        ]
-        assert len(warnings) == 2
-
-    def test_a_repaired_path_warns_again_when_it_breaks_again(
-        self, tmp_path, monkeypatch, caplog
-    ):
-        """A successful write clears the key, so a later failure is not silent."""
-        from atlas.core import prompt_risk
-
-        monkeypatch.setattr(prompt_risk, "_WRITE_FAILURE_WARNED_PATHS", set())
-        log_dir = tmp_path / "logs"
-        monkeypatch.setenv("APP_LOG_DIR", str(log_dir))
-
-        with caplog.at_level(logging.DEBUG, logger=prompt_risk.logger.name):
-            # Broken: a file sits where the log directory needs to be.
-            log_dir.write_text("")
-            self._emit(prompt_risk)
-
-            # Repaired.
-            log_dir.unlink()
-            self._emit(prompt_risk)
-            assert prompt_risk.high_risk_log_path().exists()
-
-            # Broken again, same path. A module-global ``open`` shadows the
-            # builtin the writer uses, so this fails the append without
-            # depending on filesystem permissions (tests run as root in some
-            # containers, where chmod would not bite).
-            monkeypatch.setattr(prompt_risk, "open", _raise_on_open, raising=False)
-            self._emit(prompt_risk)
-
-        warnings = [
-            r for r in caplog.records
-            if r.levelno == logging.WARNING and "high risk" in r.getMessage()
-        ]
-        assert len(warnings) == 2, (
-            "a path that is repaired and breaks again must warn again"
-        )
-
-    def test_two_unresolvable_configs_each_warn(self, monkeypatch, caplog):
-        """A resolver that raises still keys the warning by what was configured.
-
-        Keying every resolver failure on one constant would report the first
-        bad ``APP_LOG_DIR`` and silence the next one at debug.
-        """
-        from atlas.core import prompt_risk
-
-        monkeypatch.setattr(prompt_risk, "_WRITE_FAILURE_WARNED_PATHS", set())
-
-        def _raise():
-            raise RuntimeError("cannot resolve")
-
-        monkeypatch.setattr(prompt_risk, "high_risk_log_path", _raise)
-
-        with caplog.at_level(logging.DEBUG, logger=prompt_risk.logger.name):
-            monkeypatch.setenv("APP_LOG_DIR", "~nosuchuser/one")
-            self._emit(prompt_risk)
-            monkeypatch.setenv("APP_LOG_DIR", "~nosuchuser/two")
-            self._emit(prompt_risk)
-
-        warnings = [
-            r for r in caplog.records
-            if r.levelno == logging.WARNING and "high risk" in r.getMessage()
-        ]
-        assert len(warnings) == 2, "each distinct misconfiguration must warn"
-
-    def test_a_successful_write_logs_nothing(self, tmp_path, monkeypatch, caplog):
-        from atlas.core import prompt_risk
-
-        monkeypatch.setenv("APP_LOG_DIR", str(tmp_path))
-        monkeypatch.setattr(prompt_risk, "_WRITE_FAILURE_WARNED_PATHS", set())
-
-        with caplog.at_level(logging.DEBUG, logger=prompt_risk.logger.name):
-            self._emit(prompt_risk)
-
-        assert prompt_risk.high_risk_log_path().exists()
-        assert not [r for r in caplog.records if "Failed to write" in r.getMessage()]
